Add secure client to openAI clients and pass non-integration tests
