+++ src/secure-client.ts
@@ -2,12 +2,8 @@ import { Verifier } from "./verifier";
 import type { VerificationDocument } from "./verifier";
 import { TINFOIL_CONFIG } from "./config";
 import { createEncryptedBodyFetch } from "./encrypted-body-fetch";
+import { createPinnedTlsFetch } from "./pinned-tls-fetch";
 import { isRealBrowser } from "./env";
-import https from "https";
-import tls, { checkServerIdentity as tlsCheckServerIdentity } from "tls";
-import { X509Certificate, createHash } from "crypto";
-import { Readable } from "stream";
-import { ReadableStream as NodeReadableStream } from "stream/web";
 
 interface SecureClientOptions {
   baseURL?: string;
@@ -17,12 +13,12 @@ interface SecureClientOptions {
 
 export class SecureClient {
   private initPromise: Promise<void> | null = null;
+  private verificationDocument: VerificationDocument | null = null;
   private _fetch: typeof fetch | null = null;
-  private configRepo?: string;
-  private verificationDocument?: VerificationDocument;
-
-  public baseURL?: string;
-  public hpkeKeyURL?: string;
+  
+  private readonly baseURL?: string;
+  private readonly hpkeKeyURL?: string;
+  private readonly configRepo?: string;
 
   constructor(options: SecureClientOptions = {}) {
     this.baseURL = options.baseURL || TINFOIL_CONFIG.INFERENCE_BASE_URL;
@@ -44,14 +40,13 @@ export class SecureClient {
     });
 
     await verifier.verify();
-    this.verificationDocument = verifier.getVerificationDocument();
-    if (!this.verificationDocument) {
+    const doc = verifier.getVerificationDocument();
+    if (!doc) {
       throw new Error("Verification document not available after successful verification");
     }
+    this.verificationDocument = doc;
 
-    const hpkePublicKey = this.verificationDocument.enclaveMeasurement.hpkePublicKey;
-    const tlsPublicKeyFingerprint = this.verificationDocument.enclaveMeasurement.tlsPublicKeyFingerprint;
-    
+    const { hpkePublicKey, tlsPublicKeyFingerprint } = this.verificationDocument.enclaveMeasurement;
     let fetchFunction: typeof fetch;
 
     if (hpkePublicKey) {
@@ -79,7 +74,12 @@ export class SecureClient {
   }
 
   public async getVerificationDocument(): Promise<VerificationDocument> {
-    await this.ready();
+    if (!this.initPromise) {
+      await this.ready();
+    }
+    
+    await this.initPromise;
+    
     if (!this.verificationDocument) {
       throw new Error("Verification document unavailable: client not verified yet");
     }
@@ -92,111 +92,4 @@ export class SecureClient {
       return this._fetch!(input, init);
     };
   }
-}
-
-function createPinnedTlsFetch(expectedFingerprintHex: string): typeof fetch {
-  return (async (input: RequestInfo | URL, init?: RequestInit) => {
-    // Normalize URL
-    const makeURL = (value: RequestInfo | URL): URL => {
-      if (typeof value === "string") return new URL(value);
-      if (value instanceof URL) return value;
-      return new URL((value as Request).url);
-    };
-
-    const url = makeURL(input);
-    if (url.protocol !== "https:") {
-      throw new Error(`HTTP connections are not allowed. Use HTTPS. URL: ${url.toString()}`);
-    }
-
-    // Gather method and headers
-    const method = (init?.method || (input as any).method || "GET").toUpperCase();
-    const headers = new Headers(init?.headers || (input as any)?.headers || {});
-    const headerObj: Record<string, string> = {};
-    headers.forEach((v, k) => {
-      headerObj[k] = v;
-    });
-
-    // Resolve body
-    let body: any = init?.body;
-    if (!body && input instanceof Request) {
-      // If the original was a Request with a body, read it
-      try {
-        const buf = await (input as Request).arrayBuffer();
-        if (buf && (buf as ArrayBuffer).byteLength) body = Buffer.from(buf as ArrayBuffer);
-      } catch {}
-    }
-    // Convert web streams to Node streams if needed
-    if (body && typeof (body as any).getReader === "function") {
-      body = Readable.fromWeb(body as unknown as NodeReadableStream);
-    }
-    if (body instanceof ArrayBuffer) {
-      body = Buffer.from(body);
-    }
-    if (ArrayBuffer.isView(body)) {
-      body = Buffer.from(body.buffer, body.byteOffset, body.byteLength);
-    }
-
-    const requestOptions: https.RequestOptions & tls.ConnectionOptions = {
-      protocol: url.protocol,
-      hostname: url.hostname,
-      port: url.port ? Number(url.port) : 443,
-      path: `${url.pathname}${url.search}`,
-      method,
-      headers: headerObj,
-      checkServerIdentity: (host, cert): Error | undefined => {
-        const raw = (cert as any).raw as Buffer | undefined;
-        if (!raw) {
-          return new Error("Certificate raw bytes are unavailable for pinning");
-        }
-        const x509 = new X509Certificate(raw);
-        const publicKeyDer = x509.publicKey.export({ type: "spki", format: "der" });
-        const fp = createHash("sha256").update(publicKeyDer).digest("hex");
-        if (fp !== expectedFingerprintHex) {
-          return new Error(`Certificate public key fingerprint mismatch. Expected: ${expectedFingerprintHex}, Got: ${fp}`);
-        }
-        return tlsCheckServerIdentity(host, cert);
-      },
-    };
-
-    const { signal } = init || {};
-
-    const res = await new Promise<import("http").IncomingMessage>((resolve, reject) => {
-      const req = https.request(requestOptions, resolve);
-      req.on("error", reject);
-      if (signal) {
-        if ((signal as AbortSignal).aborted) {
-          req.destroy(new Error("Request aborted"));
-          return;
-        }
-        (signal as AbortSignal).addEventListener("abort", () => req.destroy(new Error("Request aborted")));
-      }
-      if (body === undefined || body === null) {
-        req.end();
-      } else if (typeof body === "string" || Buffer.isBuffer(body) || ArrayBuffer.isView(body)) {
-        req.end(body as any);
-      } else if (typeof (body as any).pipe === "function") {
-        (body as any).pipe(req);
-      } else {
-        // Fallback: try to serialize objects
-        req.end(String(body));
-      }
-    });
-
-    const responseHeaders = new Headers();
-    for (const [k, v] of Object.entries(res.headers)) {
-      if (Array.isArray(v)) {
-        v.forEach(item => responseHeaders.append(k, item));
-      } else if (v != null) {
-        responseHeaders.set(k, String(v));
-      }
-    }
-
-    // Convert Node stream to Web ReadableStream
-    const webStream = Readable.toWeb(res as unknown as import("stream").Readable) as unknown as ReadableStream;
-    return new Response(webStream, {
-      status: res.statusCode || 0,
-      statusText: res.statusMessage || "",
-      headers: responseHeaders,
-    });
-  }) as typeof fetch;
 }
